chore: split zustand store
