+++ apps/web/src/components/Common/Auth/ConnectWalletButton.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@components/UIElements/Button'
 import Tooltip from '@components/UIElements/Tooltip'
-import useAppStore from '@lib/store'
+import useChannelStore from '@lib/store/channel'
 import usePersistStore from '@lib/store/persist'
 import { useConnectModal } from '@rainbow-me/rainbowkit'
 import React from 'react'
@@ -19,7 +19,7 @@ type Props = {
 
 const ConnectWalletButton = ({ handleSign, signing }: Props) => {
   const selectedChannelId = usePersistStore((state) => state.selectedChannelId)
-  const selectedChannel = useAppStore((state) => state.selectedChannel)
+  const selectedChannel = useChannelStore((state) => state.selectedChannel)
 
   const { connector, isConnected } = useAccount()
   const { switchNetwork } = useSwitchNetwork({
