+++ apps/web/src/components/Settings/DangerZone.tsx
@@ -1,7 +1,7 @@
 import { LENSHUB_PROXY_ABI } from '@abis/LensHubProxy'
 import IsVerified from '@components/Common/IsVerified'
 import { Button } from '@components/UIElements/Button'
-import useAppStore from '@lib/store'
+import useChannelStore from '@lib/store/channel'
 import { signOut } from '@lib/store/persist'
 import { utils } from 'ethers'
 import type { CreateBurnProfileBroadcastItemResult } from 'lens'
@@ -21,9 +21,10 @@ import {
 } from 'wagmi'
 
 const DangerZone = () => {
-  const selectedChannel = useAppStore((state) => state.selectedChannel)
   const [loading, setLoading] = useState(false)
   const [txnHash, setTxnHash] = useState<`0x${string}`>()
+  const selectedChannel = useChannelStore((state) => state.selectedChannel)
+
   const { signTypedDataAsync } = useSignTypedData({
     onError(error) {
       toast.error(error?.message)
