+++ apps/web/src/lib/store/persist.ts
@@ -8,29 +8,19 @@ type Tokens = {
 }
 
 interface AppPerisistState {
-  accessToken: Tokens['accessToken']
-  refreshToken: Tokens['refreshToken']
-  selectedChannelId: string | null
   sidebarCollapsed: boolean
   latestNotificationId: string
   queuedVideos: QueuedVideoType[]
   queuedComments: QueuedCommentType[]
   setLatestNotificationId: (id: string) => void
   setSidebarCollapsed: (collapsed: boolean) => void
-  setSelectedChannelId: (id: string | null) => void
   setQueuedComments: (queuedComments: QueuedCommentType[]) => void
   setQueuedVideos: (queuedVideos: QueuedVideoType[]) => void
-  signIn: (tokens: { accessToken: string; refreshToken: string }) => void
-  signOut: () => void
-  hydrateAuthTokens: () => Tokens
 }
 
 export const usePersistStore = create(
   persist<AppPerisistState>(
     (set, get) => ({
-      accessToken: null,
-      refreshToken: null,
-      selectedChannelId: null,
       sidebarCollapsed: true,
       latestNotificationId: '',
       queuedComments: [],
@@ -39,17 +29,7 @@ export const usePersistStore = create(
       setQueuedComments: (queuedComments) => set({ queuedComments }),
       setSidebarCollapsed: (sidebarCollapsed) => set({ sidebarCollapsed }),
       setLatestNotificationId: (latestNotificationId) =>
-        set({ latestNotificationId }),
-      setSelectedChannelId: (id) => set({ selectedChannelId: id }),
-      signIn: ({ accessToken, refreshToken }) =>
-        set({ accessToken, refreshToken }),
-      signOut: () => localStorage.removeItem('lenstube.store'),
-      hydrateAuthTokens: () => {
-        return {
-          accessToken: get().accessToken,
-          refreshToken: get().refreshToken
-        }
-      }
+        set({ latestNotificationId })
     }),
     {
       name: 'lenstube.store'
@@ -58,9 +38,3 @@ export const usePersistStore = create(
 )
 
 export default usePersistStore
-
-export const signIn = (tokens: { accessToken: string; refreshToken: string }) =>
-  usePersistStore.getState().signIn(tokens)
-export const signOut = () => usePersistStore.getState().signOut()
-export const hydrateAuthTokens = () =>
-  usePersistStore.getState().hydrateAuthTokens()
