+++ apps/web/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import '@rainbow-me/rainbowkit/styles.css'
 import 'tippy.js/dist/tippy.css'
 
 import FullPageLoader from '@components/Common/FullPageLoader'
-import usePersistStore from '@lib/store/persist'
+import useAuthPersistStore from '@lib/store/auth'
 import type { AppProps } from 'next/app'
 import { useRouter } from 'next/router'
 import React, { lazy, Suspense, useEffect } from 'react'
@@ -13,8 +13,10 @@ const Providers = lazy(() => import('../components/Common/Providers'))
 const Layout = lazy(() => import('../components/Common/Layout'))
 
 const App = ({ Component, pageProps }: AppProps) => {
-  const selectedChannelId = usePersistStore((state) => state.selectedChannelId)
   const { pathname, replace, asPath } = useRouter()
+  const selectedChannelId = useAuthPersistStore(
+    (state) => state.selectedChannelId
+  )
 
   useEffect(() => {
     if (!selectedChannelId && AUTH_ROUTES.includes(pathname)) {
